requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-database-url==2.3.0
dj-rest-auth==7.0.0
Django==5.1.2
django-filter==24.3
django-oauth-plus==2.2.9
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.1
httplib2==0.22.0
idna==3.10
Markdown==3.7
oauth2==1.9.0.post1
oauthlib==3.2.2
psycopg==3.2.3
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
